refactor(search): split the tika extractor and its tests into per-facet files
