+++ services/search/pkg/content/tika_image_test.go
@@ -1,0 +1,23 @@
+package content
+
+import (
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+	libregraph "github.com/opencloud-eu/libre-graph-api-go"
+)
+
+var _ = Describe("getImage", func() {
+	It("maps the image dimensions to the image facet", func() {
+		image := Tika{}.getImage(map[string][]string{
+			"tiff:ImageWidth":  {"100"},
+			"tiff:ImageLength": {"200"},
+		})
+		Expect(image).ToNot(BeNil())
+		Expect(image.Width).To(Equal(libregraph.PtrInt32(100)))
+		Expect(image.Height).To(Equal(libregraph.PtrInt32(200)))
+	})
+
+	It("returns nil when no image metadata is present", func() {
+		Expect(Tika{}.getImage(map[string][]string{})).To(BeNil())
+	})
+})
